Shorten test name to suit Windows MAX_PATH limit
